requirements.txt
discord.py
requests
pandas
ta
python-dotenv